Extract normals-visualizer builders into engine/src/normalsVisualizer.ts
Moves createNormalsVisualizer, createComputedNormalsVisualizer, and
createPointCloudNormalsVisualizer out of main.ts as pure functions (they
had no dependency on instance state).
